+++ app.py
@@ -1,11 +1,7 @@
 from flask import Flask, jsonify
 from flask import request
-import settings
-settings.db_conf()
-settings.dbhost = 'localhost'
-settings.dbuser = 'root'
-settings.dbpassword = ''
-settings.dbname = 's1'
+
+import pythoncrawler.settings
 
 app = Flask(__name__)
 
@@ -26,7 +22,7 @@ def index():
         </html>
     '''
 
-# Database 
+# Database
 import db_connection
 conn = db_connection.db_conn()
 a = conn.cursor()
