+++ pythoncrawler/regex.py
@@ -1,0 +1,3 @@
+def regex():
+    global domain_regex
+    domain_regex = '(\/\/)(?:[\w-]+\.)*([\w-]{1,63})(?:\.(?:\w{3}|\w{2}))'
